Delet all over layers back to 1-03-2009
I am  satooshi  nakmoto  and  i  never  approved  any  layers  and  to be  part of  is  and  was  to  give the  idea  of  p2p  be posiblem many  have said anyone  can  upgrade  or update  the  code  that is  wrong  no one   can update anything or change anything or authorized to give authorization to access bitcoin including coinbass which i officaly is cossing all my walllets of the coinpass which will return to blockchain to be secure will all my funds and finace and. BtC bank also take back whats mine.